pyproject.toml
[tool.poetry]
name = "ecdsa"
version = "0.1.0"
description = "ECDSA study example"
authors = ["Maicol K. Oliveira <maicolkaiser.oliveira@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pysha3 = "^1.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
